better support for embedded scaffolds. each embedded scaffold gets a unique id which is used to separate its session storage.
an embedded scaffold is any scaffold created via the "render :active_scaffold" signature. the embedded id ('eid') is created as the md5 of the controller doing the 
embedding, the controller serving the scaffold, and any parameters passed to the embedded scaffold. this means that if someone embeds the Users scaffold with the 
same constraints (e.g. user_group_id=5) from two different controllers, those scaffolds will keep state independently.

constraints are now stored in the session storage. these constraints will be used in the near future to intelligently remove columns from appearing in the list or 
update/create forms, and will be automatically inserted into new models.



git-svn-id: http://activescaffold.googlecode.com/svn/trunk@12 561dde7e-7729-0410-be8e-ef83869d6c7d