big changeset for nested scaffolds. this changeset beefs up the intelligence of constraints for embedded scaffolds. constraints may now be column/association names, and if a constraint is an association name then the constraint value will be assumed an ID. this should make nested scaffolds work with habtm and has_many both.
hopefully this will fix some other odd issues around nested scaffolds, too. or cause more ... sigh.

in order for nested scaffolds to use the new changeset, they now depend on the new association.reverse property. this property specifies the "reverse association", 
which normally is not available. we try and guess the reverse association by standard convention, but if we fail then nested scaffolds will not work until the 
developer specifies the reverse association himself.

what is a reverse association? for example:

  class User < ActiveRecord::Base
    # the reverse association of :user_group is ":users"
    belongs_to :user_group
  end

  class UserGroup < ActiveRecord::Base
    # the reverse association of :users is ":user_group"
    has_many :users
  end




git-svn-id: http://activescaffold.googlecode.com/svn/trunk@223 561dde7e-7729-0410-be8e-ef83869d6c7d