refactoring the validation/saving for associated records. i'm using a variation on dave ringoen's approach to figure out which records need to be saved, which was something like a dirty flag. one difference in my take on the solution is that the dirty flag is automatically reset when the record is saved.
i started to try and deal with validates_presence_of in this changeset, but it got too complicated and we decided to take another look at the problem. it may be 
that the rails docs are wrong when they recommend using 'validates_presence_of :foreign_key' instead of 'validates_presence_of :association_name'. the latter works 
even when the associated record is new and there is no id yet, and because rails automatically saves belongs_to associations before the current record there 
shouldn't be any actual database problems (unless the associated record is invalid and doesn't get saved, which can be solved with 'validates_associated').



git-svn-id: http://activescaffold.googlecode.com/svn/trunk@433 561dde7e-7729-0410-be8e-ef83869d6c7d