register_cancel_hooks fix for Firefox 3
as noted in
  http://blog.caboo.se/articles/2008/3/21/safari-3-1-breaks-your-app
Firefox 3 and Safari 3.1 both implement their own getElementsByClassName which breaks register_cancel_hooks.

this is easy to get around by replacing getElementsByClassName() with Prototype's select()

this has been tested by me on FF2, FF3, IE7.
according to
  http://groups.google.com/group/activescaffold/browse_thread/thread/d37afa3f2ac687f2/b34ebf62d767bee3
it has also been tested on Safari 3.1

thanks to everyone on the above threads who contributed variations on this patch.