Avoid DB errors when clearing a value in an inline edit
If a database column is declared with a not-NULL constraint, then
do_update_column should not set the attribute to nil in an effort
to clear it.  Although doing so works with MySQL, it fails with
PostgreSQL.

Instead, if no attribute value is specified and the record's column
cannot be NULL, the attribute value is set to the column's default
value.