fix generate_delete_helper for ruby 1.9
This method checks for delete_{field} being defined be looking at
kalss.instance_methods.  In ruby 1.9 this returns an array of symbols.
In ruby 1.8.7 it was strings.  So the check for the method being
defined always fails now.  I changed it to look up the symbol rather
than the string