[TinyMCE] Load editor settings from config/tinymce.yml when js_framework is :jquery
When the js_framework is set to jQuery the bridge did not load the
settings from config/tinymce.yml as it did in the previous
versions. This patch fixes that regression by calling the
tinymce-rails helper functions for creating tinymce instances when the
framework is set to jquery, which does the appropriate loading from
config/tinymce.yml.