Override the RenderingHelper render method
When a partial calls `render :super`, the view_paths for the second
render attempt are not available when action view attempts to render
using `view_renderer.render` because view_renderer contains a memoized
ActionView::Renderer.

Fix this by explictly instantiating a new ActionView::Renderer instance
and call render on that.  There may be a better solution than this.