Add hide_column_if column attr_accessor
Add a new attr_accessor to the DataStructures::Column module to allow
for calling a dynamic Proc to determine whether a form column should be
shown or not.

When hidden, render a dl element that can still be re-rendered by the
update_columns routine so we will re-evaluate the proc whenever the
column is re-rendered.

The value can be set a Proc in the scaffold controller as follows:
config.columns[:my_column].hide_column_if = Proc.new { |record, column, scope| record.should_be_hidden? }

OR to always hide, use:
config.columns[:my_column].hide_column_if = true

This is useful when you want the column to be accessible via API but do
not want it to be visible in the HTML form.