Fix label lookup process when controller doesn't match model
When attempting to find the label for a controller rendered with the
`render active_scaffold: 'path/to/controller'` helper, use the controller
path to infer the scaffold controller, rather than parsing a model from
the controller path, and then attempting to find the active scaffold config
for that class.

Inferring the controller directly is preferable because there are cases
where the controller name does not match the underlying model.  In this
case when we parse the model from the controller name, we end up trying
to constantize the wrong model class name, causing a NameError.

If we can't find the config from the controller name for some reason,
fall back to the old behavior of pulling the model from the controller
name and using active_scaffold_config_for(model)

Fixes activescaffold/active_scaffold/#711