Add support for dynamic form descriptions
allow a scaffold controller to define a column's description attribute
as a Proc, such that the description content can change depending on the
current record.  This is useful when another field in the form (such as
a "type" dropdown) can change the context for the record enough that the
description should now reflect different information.  This would
typically be used in conjunction with a column that uses update_columns
to re-render other fields when it changes.