support refreshing fields which are not at the same level
So the values can be column names, or hashes, with association as keys, and array of column names of the associated model as values.

Also supports :__root__ as key, to refresh fields on the form, so changing a column in a subform can trigger refreshing a column in the form, or another subform with nested hashes.